refactor(core): organize domain modules
