+++ crates/omniinfer-core/src/backend/mod.rs
@@ -1,0 +1,5 @@
+pub mod args;
+mod detection;
+pub mod profiles;
+pub mod registry;
+mod templates;
