+++ crates/omniinfer-core/src/backend/registry.rs
@@ -4,6 +4,10 @@ use std::path::{Path, PathBuf};
 use serde::Serialize;
 use serde_json::{Value, json};
 
+use super::detection::{embedded_module_exists, is_hardware_compatible};
+#[cfg(test)]
+use super::detection::{gpu_backend_ids, output_mentions_amd_gpu};
+use super::templates::backend_templates;
 use crate::{config, local_state, paths};
 
 const LLAMA_CPP_CACHE_RAM_MIB: &str = "8192";
@@ -475,195 +479,6 @@ fn split_extra_args(text: &str) -> Vec<String> {
     args
 }
 
-fn is_hardware_compatible(host: HostInfo, spec: &BackendSpec) -> bool {
-    let caps = spec
-        .capabilities
-        .iter()
-        .map(String::as_str)
-        .collect::<Vec<_>>();
-    if caps.contains(&"arm64") && !is_arm64(host.machine) {
-        return false;
-    }
-    if caps.contains(&"x64") && !is_x86_64(host.machine) {
-        return false;
-    }
-    if caps.contains(&"s390x") && host.machine != "s390x" {
-        return false;
-    }
-    if caps.contains(&"openvino") || caps.contains(&"eagle3") {
-        return spec.binary_exists();
-    }
-    if !gpu_backend_ids(host).contains(&spec.id.as_str()) {
-        return true;
-    }
-    if caps.contains(&"cuda") {
-        return cuda_detected();
-    }
-    if caps.contains(&"rocm") || caps.contains(&"hip") {
-        return rocm_detected(host);
-    }
-    if caps.contains(&"metal") {
-        return host.system == HostSystem::Mac || host.system == HostSystem::Ios;
-    }
-    if caps.contains(&"vulkan") {
-        return vulkan_detected();
-    }
-    spec.binary_exists()
-}
-
-fn is_arm64(machine: &str) -> bool {
-    machine.eq_ignore_ascii_case("aarch64") || machine.eq_ignore_ascii_case("arm64")
-}
-
-fn is_x86_64(machine: &str) -> bool {
-    machine.eq_ignore_ascii_case("x86_64") || machine.eq_ignore_ascii_case("amd64")
-}
-
-fn cuda_detected() -> bool {
-    std::process::Command::new("nvidia-smi")
-        .args(["--query-gpu=index", "--format=csv,noheader,nounits"])
-        .output()
-        .map(|output| output.status.success() && !output.stdout.is_empty())
-        .unwrap_or(false)
-}
-
-fn rocm_detected(host: HostInfo) -> bool {
-    if host.system == HostSystem::Windows {
-        return windows_amd_gpu_detected();
-    }
-    std::process::Command::new("rocm-smi")
-        .arg("--showmeminfo")
-        .output()
-        .map(|output| output.status.success())
-        .unwrap_or(false)
-}
-
-fn windows_amd_gpu_detected() -> bool {
-    let mut powershell = std::process::Command::new("powershell.exe");
-    powershell.args([
-        "-NoProfile",
-        "-NonInteractive",
-        "-Command",
-        "Get-CimInstance Win32_VideoController | ForEach-Object Name",
-    ]);
-    hide_child_window(&mut powershell);
-    if powershell
-        .output()
-        .is_ok_and(|output| output.status.success() && output_mentions_amd_gpu(&output.stdout))
-    {
-        return true;
-    }
-
-    let mut registry = std::process::Command::new("reg.exe");
-    registry.args([
-        "query",
-        r"HKLM\SYSTEM\CurrentControlSet\Control\Class\{4d36e968-e325-11ce-bfc1-08002be10318}",
-        "/s",
-        "/v",
-        "DriverDesc",
-    ]);
-    hide_child_window(&mut registry);
-    registry
-        .output()
-        .is_ok_and(|output| output.status.success() && output_mentions_amd_gpu(&output.stdout))
-}
-
-fn output_mentions_amd_gpu(output: &[u8]) -> bool {
-    let names = String::from_utf8_lossy(output).to_ascii_lowercase();
-    names.contains("radeon")
-        || names
-            .split(|character: char| !character.is_ascii_alphanumeric())
-            .any(|token| token == "amd")
-}
-
-fn hide_child_window(command: &mut std::process::Command) {
-    #[cfg(windows)]
-    {
-        use std::os::windows::process::CommandExt;
-        const CREATE_NO_WINDOW: u32 = 0x0800_0000;
-        command.creation_flags(CREATE_NO_WINDOW);
-    }
-    #[cfg(not(windows))]
-    {
-        let _ = command;
-    }
-}
-
-fn vulkan_detected() -> bool {
-    std::process::Command::new("vulkaninfo")
-        .arg("--summary")
-        .output()
-        .map(|output| output.status.success())
-        .unwrap_or(false)
-}
-
-fn embedded_module_exists(runtime_dir: &Path, module_name: &str) -> bool {
-    embedded_site_roots(runtime_dir)
-        .iter()
-        .any(|root| module_path_exists(root, module_name))
-}
-
-fn embedded_site_roots(runtime_dir: &Path) -> Vec<PathBuf> {
-    let mut roots = Vec::new();
-    let bases = [runtime_dir.to_path_buf(), runtime_dir.join("venv")];
-    for base in bases {
-        for candidate in [
-            base.join("Lib").join("site-packages"),
-            base.join("lib").join("site-packages"),
-        ] {
-            if candidate.is_dir() && !roots.contains(&candidate) {
-                roots.push(candidate);
-            }
-        }
-        let pattern_root = base.join("lib");
-        if let Ok(entries) = std::fs::read_dir(pattern_root) {
-            for entry in entries.flatten() {
-                let path = entry.path();
-                let Some(name) = path.file_name().and_then(|name| name.to_str()) else {
-                    continue;
-                };
-                if !name.starts_with("python") {
-                    continue;
-                }
-                for site_name in ["site-packages", "dist-packages"] {
-                    let candidate = path.join(site_name);
-                    if candidate.is_dir() && !roots.contains(&candidate) {
-                        roots.push(candidate);
-                    }
-                }
-            }
-        }
-    }
-    roots
-}
-
-fn module_path_exists(site_root: &Path, module_name: &str) -> bool {
-    let module_path = module_name
-        .split('.')
-        .fold(site_root.to_path_buf(), |path, item| path.join(item));
-    if module_path.is_dir() || module_path.with_extension("py").is_file() {
-        return true;
-    }
-    let Some(parent) = module_path.parent() else {
-        return false;
-    };
-    let Some(name) = module_path.file_name().and_then(|name| name.to_str()) else {
-        return false;
-    };
-    std::fs::read_dir(parent)
-        .map(|entries| {
-            entries.flatten().any(|entry| {
-                let path = entry.path();
-                path.file_stem().and_then(|stem| stem.to_str()) == Some(name)
-                    && matches!(
-                        path.extension().and_then(|ext| ext.to_str()),
-                        Some("so" | "pyd" | "dll" | "dylib")
-                    )
-            })
-        })
-        .unwrap_or(false)
-}
-
 fn recommended_backend(rows: &[Value]) -> Option<String> {
     rows.iter()
         .filter(|row| backend_payload_has_capability(row, "chat"))
@@ -702,40 +517,6 @@ fn backend_payload_has_capability(row: &Value, wanted: &str) -> bool {
         .flatten()
         .filter_map(Value::as_str)
         .any(|capability| capability == wanted)
-}
-
-fn gpu_backend_ids(host: HostInfo) -> &'static [&'static str] {
-    match host.system {
-        HostSystem::Linux => &[
-            "llama.cpp-linux-cuda",
-            "llama.cpp-linux-rocm",
-            "llama.cpp-linux-vulkan",
-            "omniinfer-native-linux",
-            "ik_llama.cpp-linux-cuda",
-            "vllm-linux-cuda",
-            "vla.cpp-linux-cuda",
-        ],
-        HostSystem::Windows => &[
-            "llama.cpp-cuda",
-            "llama.cpp-vulkan",
-            "llama.cpp-sycl",
-            "llama.cpp-hip",
-            "ik_llama.cpp-cuda",
-            "vllm-wsl2-cuda",
-            "vllm-wsl2-rocm",
-        ],
-        _ => &[],
-    }
-}
-
-fn backend_templates(host: HostInfo) -> &'static [BackendTemplate] {
-    match host.system {
-        HostSystem::Linux => LINUX_TEMPLATES,
-        HostSystem::Windows => WINDOWS_TEMPLATES,
-        HostSystem::Mac => MAC_TEMPLATES,
-        HostSystem::Android => ANDROID_TEMPLATES,
-        HostSystem::Ios => IOS_TEMPLATES,
-    }
 }
 
 fn env_value(name: &str) -> Option<String> {
@@ -777,528 +558,6 @@ fn home_dir() -> Option<PathBuf> {
         .or_else(|| std::env::var_os("USERPROFILE"))
         .map(PathBuf::from)
 }
-
-const fn template(
-    id: &'static str,
-    label: &'static str,
-    family: &'static str,
-    runtime_dir_name: &'static str,
-    launcher_name: Option<&'static str>,
-    description: &'static str,
-    capabilities: &'static [&'static str],
-    env_prefix: &'static str,
-) -> BackendTemplate {
-    BackendTemplate {
-        id,
-        label,
-        family,
-        runtime_dir_name,
-        launcher_name,
-        description,
-        capabilities,
-        env_prefix,
-        default_ngl: None,
-        default_extra_args: &[],
-        fallback_runtime_dir_names: &[],
-        runtime_mode: "external_server",
-        model_artifact: "file",
-        supports_mmproj: true,
-        supports_ctx_size: true,
-        python_modules: &[],
-        external_server_protocol: Some("llama.cpp-server"),
-        log_file_name: "runtime.log",
-    }
-}
-
-const LINUX_TEMPLATES: &[BackendTemplate] = &[
-    template(
-        "llama.cpp-linux",
-        "llama.cpp Linux",
-        "llama.cpp",
-        "llama.cpp-linux",
-        Some("llama-server"),
-        "llama.cpp Linux CPU backend managed by OmniInfer",
-        &["chat", "vision", "stream", "cpu", "linux"],
-        "OMNIINFER_LLAMA_CPP_LINUX",
-    ),
-    BackendTemplate {
-        default_ngl: Some("999"),
-        ..template(
-            "llama.cpp-linux-cuda",
-            "llama.cpp Linux CUDA",
-            "llama.cpp",
-            "llama.cpp-linux-cuda",
-            Some("llama-server"),
-            "llama.cpp Linux CUDA backend managed by OmniInfer",
-            &["chat", "vision", "stream", "gpu", "cuda", "linux"],
-            "OMNIINFER_LLAMA_CPP_LINUX_CUDA",
-        )
-    },
-    BackendTemplate {
-        default_ngl: Some("999"),
-        fallback_runtime_dir_names: &["llama.cpp-linux-ROCm"],
-        ..template(
-            "llama.cpp-linux-rocm",
-            "llama.cpp Linux ROCm",
-            "llama.cpp",
-            "llama.cpp-linux-rocm",
-            Some("llama-server"),
-            "llama.cpp Linux ROCm backend managed by OmniInfer",
-            &["chat", "vision", "stream", "gpu", "rocm", "linux"],
-            "OMNIINFER_LLAMA_CPP_LINUX_ROCM",
-        )
-    },
-    BackendTemplate {
-        default_ngl: Some("999"),
-        ..template(
-            "llama.cpp-linux-vulkan",
-            "llama.cpp Linux Vulkan",
-            "llama.cpp",
-            "llama.cpp-linux-vulkan",
-            Some("llama-server"),
-            "llama.cpp Linux Vulkan backend managed by OmniInfer",
-            &["chat", "vision", "stream", "gpu", "vulkan", "linux"],
-            "OMNIINFER_LLAMA_CPP_LINUX_VULKAN",
-        )
-    },
-    template(
-        "llama.cpp-linux-s390x",
-        "llama.cpp Linux s390x",
-        "llama.cpp",
-        "llama.cpp-linux-s390x",
-        Some("llama-server"),
-        "llama.cpp Linux s390x CPU backend managed by OmniInfer",
-        &["chat", "vision", "stream", "cpu", "linux", "s390x"],
-        "OMNIINFER_LLAMA_CPP_LINUX_S390X",
-    ),
-    BackendTemplate {
-        default_ngl: Some("999"),
-        ..template(
-            "omniinfer-native-linux",
-            "OmniInfer Native Linux (EAGLE3)",
-            "llama.cpp",
-            "omniinfer-native-linux",
-            Some("llama-server"),
-            "OmniInfer Native Linux backend with EAGLE3 speculative decoding",
-            &[
-                "chat", "vision", "stream", "gpu", "vulkan", "linux", "eagle3",
-            ],
-            "OMNIINFER_NATIVE_LINUX",
-        )
-    },
-    template(
-        "llama.cpp-linux-openvino",
-        "llama.cpp Linux OpenVINO",
-        "llama.cpp",
-        "llama.cpp-linux-openvino",
-        Some("llama-server"),
-        "llama.cpp Linux OpenVINO backend managed by OmniInfer",
-        &["chat", "vision", "stream", "linux", "openvino", "intel"],
-        "OMNIINFER_LLAMA_CPP_LINUX_OPENVINO",
-    ),
-    BackendTemplate {
-        default_extra_args: &["--jinja"],
-        ..template(
-            "ik_llama.cpp-linux",
-            "ik_llama.cpp Linux",
-            "llama.cpp",
-            "ik_llama.cpp-linux",
-            Some("llama-server"),
-            "ik_llama.cpp Linux CPU backend managed by OmniInfer",
-            &["chat", "vision", "stream", "cpu", "linux"],
-            "OMNIINFER_IK_LLAMA_CPP_LINUX",
-        )
-    },
-    BackendTemplate {
-        default_ngl: Some("999"),
-        default_extra_args: &["--jinja"],
-        ..template(
-            "ik_llama.cpp-linux-cuda",
-            "ik_llama.cpp Linux CUDA",
-            "llama.cpp",
-            "ik_llama.cpp-linux-cuda",
-            Some("llama-server"),
-            "ik_llama.cpp Linux CUDA backend managed by OmniInfer",
-            &["chat", "vision", "stream", "gpu", "cuda", "linux"],
-            "OMNIINFER_IK_LLAMA_CPP_LINUX_CUDA",
-        )
-    },
-    BackendTemplate {
-        runtime_mode: "embedded",
-        model_artifact: "path",
-        supports_mmproj: false,
-        supports_ctx_size: false,
-        python_modules: &["MNN", "MNN.llm", "MNN.cv"],
-        external_server_protocol: None,
-        ..template(
-            "mnn-linux",
-            "MNN Linux",
-            "mnn",
-            "mnn-linux",
-            None,
-            "Embedded MNN LLM/VLM backend managed directly by OmniInfer on Linux",
-            &[
-                "chat", "vision", "stream", "cpu", "linux", "embedded", "mnn",
-            ],
-            "OMNIINFER_MNN_LINUX",
-        )
-    },
-    BackendTemplate {
-        model_artifact: "reference",
-        supports_mmproj: false,
-        external_server_protocol: Some("vllm-openai-server"),
-        log_file_name: "vllm-server.log",
-        ..template(
-            "vllm-linux-cuda",
-            "vLLM Linux CUDA",
-            "vllm",
-            "vllm-linux-cuda",
-            Some("vllm"),
-            "vLLM OpenAI-compatible server backend managed by OmniInfer on Linux CUDA",
-            &[
-                "chat",
-                "stream",
-                "gpu",
-                "cuda",
-                "linux",
-                "openai-compatible",
-            ],
-            "OMNIINFER_VLLM_LINUX_CUDA",
-        )
-    },
-    BackendTemplate {
-        model_artifact: "vla-artifact",
-        supports_ctx_size: false,
-        external_server_protocol: Some("vla.cpp-zmq-server"),
-        log_file_name: "vla-server.log",
-        ..template(
-            "vla.cpp-linux",
-            "vla.cpp Linux",
-            "vla.cpp",
-            "vla.cpp-linux",
-            Some("vla-server"),
-            "vla.cpp ZeroMQ/protobuf VLA action server managed by OmniInfer on Linux CPU",
-            &[
-                "vision", "action", "robotics", "cpu", "linux", "zeromq", "protobuf",
-            ],
-            "OMNIINFER_VLA_CPP_LINUX",
-        )
-    },
-    BackendTemplate {
-        model_artifact: "vla-artifact",
-        supports_ctx_size: false,
-        external_server_protocol: Some("vla.cpp-zmq-server"),
-        log_file_name: "vla-server.log",
-        ..template(
-            "vla.cpp-linux-cuda",
-            "vla.cpp Linux CUDA",
-            "vla.cpp",
-            "vla.cpp-linux-cuda",
-            Some("vla-server"),
-            "vla.cpp ZeroMQ/protobuf VLA action server managed by OmniInfer on Linux CUDA",
-            &[
-                "vision", "action", "robotics", "gpu", "cuda", "linux", "zeromq", "protobuf",
-            ],
-            "OMNIINFER_VLA_CPP_LINUX_CUDA",
-        )
-    },
-];
-
-const WINDOWS_TEMPLATES: &[BackendTemplate] = &[
-    template(
-        "llama.cpp-cpu",
-        "llama.cpp cpu",
-        "llama.cpp",
-        "llama.cpp-cpu",
-        Some("llama-server.exe"),
-        "llama.cpp CPU backend managed by OmniInfer",
-        &["chat", "vision", "stream", "cpu"],
-        "OMNIINFER_LLAMA_CPP_CPU",
-    ),
-    BackendTemplate {
-        default_ngl: Some("999"),
-        ..template(
-            "llama.cpp-cuda",
-            "llama.cpp CUDA",
-            "llama.cpp",
-            "llama.cpp-cuda",
-            Some("llama-server.exe"),
-            "llama.cpp CUDA backend managed by OmniInfer",
-            &["chat", "vision", "stream", "gpu", "cuda"],
-            "OMNIINFER_LLAMA_CPP_CUDA",
-        )
-    },
-    BackendTemplate {
-        default_ngl: Some("999"),
-        ..template(
-            "llama.cpp-vulkan",
-            "llama.cpp Vulkan",
-            "llama.cpp",
-            "llama.cpp-vulkan",
-            Some("llama-server.exe"),
-            "llama.cpp Vulkan backend managed by OmniInfer",
-            &["chat", "vision", "stream", "gpu", "vulkan"],
-            "OMNIINFER_LLAMA_CPP_VULKAN",
-        )
-    },
-    template(
-        "llama.cpp-windows-arm64",
-        "llama.cpp Windows arm64",
-        "llama.cpp",
-        "llama.cpp-windows-arm64",
-        Some("llama-server.exe"),
-        "llama.cpp Windows arm64 CPU backend managed by OmniInfer",
-        &["chat", "vision", "stream", "cpu", "windows", "arm64"],
-        "OMNIINFER_LLAMA_CPP_WINDOWS_ARM64",
-    ),
-    BackendTemplate {
-        default_ngl: Some("999"),
-        ..template(
-            "llama.cpp-sycl",
-            "llama.cpp SYCL",
-            "llama.cpp",
-            "llama.cpp-sycl",
-            Some("llama-server.exe"),
-            "llama.cpp Windows SYCL backend managed by OmniInfer",
-            &[
-                "chat", "vision", "stream", "gpu", "sycl", "intel", "windows",
-            ],
-            "OMNIINFER_LLAMA_CPP_SYCL",
-        )
-    },
-    BackendTemplate {
-        default_ngl: Some("999"),
-        ..template(
-            "llama.cpp-hip",
-            "llama.cpp HIP",
-            "llama.cpp",
-            "llama.cpp-hip",
-            Some("llama-server.exe"),
-            "llama.cpp Windows HIP backend managed by OmniInfer",
-            &["chat", "vision", "stream", "gpu", "hip", "amd", "windows"],
-            "OMNIINFER_LLAMA_CPP_HIP",
-        )
-    },
-    BackendTemplate {
-        default_extra_args: &["--jinja"],
-        ..template(
-            "ik_llama.cpp-cpu",
-            "ik_llama.cpp CPU",
-            "llama.cpp",
-            "ik_llama.cpp-cpu",
-            Some("llama-server.exe"),
-            "ik_llama.cpp CPU backend managed by OmniInfer",
-            &["chat", "vision", "stream", "cpu"],
-            "OMNIINFER_IK_LLAMA_CPP_CPU",
-        )
-    },
-    BackendTemplate {
-        default_ngl: Some("999"),
-        default_extra_args: &["--jinja"],
-        ..template(
-            "ik_llama.cpp-cuda",
-            "ik_llama.cpp CUDA",
-            "llama.cpp",
-            "ik_llama.cpp-cuda",
-            Some("llama-server.exe"),
-            "ik_llama.cpp CUDA backend managed by OmniInfer",
-            &["chat", "vision", "stream", "gpu", "cuda"],
-            "OMNIINFER_IK_LLAMA_CPP_CUDA",
-        )
-    },
-    BackendTemplate {
-        model_artifact: "reference",
-        supports_mmproj: false,
-        external_server_protocol: Some("vllm-wsl2-openai-server"),
-        log_file_name: "vllm-wsl2-server.log",
-        ..template(
-            "vllm-wsl2-cuda",
-            "vLLM WSL2 CUDA",
-            "vllm",
-            "vllm-wsl2-cuda",
-            Some("vllm-wsl2.json"),
-            "Official vLLM Linux CUDA runtime managed by OmniInfer through WSL2",
-            &[
-                "chat",
-                "stream",
-                "gpu",
-                "cuda",
-                "windows",
-                "wsl2",
-                "openai-compatible",
-            ],
-            "OMNIINFER_VLLM_WSL2_CUDA",
-        )
-    },
-    BackendTemplate {
-        model_artifact: "reference",
-        supports_mmproj: false,
-        external_server_protocol: Some("vllm-wsl2-openai-server"),
-        log_file_name: "vllm-wsl2-rocm-server.log",
-        ..template(
-            "vllm-wsl2-rocm",
-            "vLLM WSL2 ROCm",
-            "vllm",
-            "vllm-wsl2-rocm",
-            Some("vllm-wsl2.json"),
-            "Official vLLM Linux ROCm runtime for Ryzen AI managed by OmniInfer through WSL2",
-            &[
-                "chat",
-                "stream",
-                "gpu",
-                "rocm",
-                "amd",
-                "windows",
-                "wsl2",
-                "openai-compatible",
-            ],
-            "OMNIINFER_VLLM_WSL2_ROCM",
-        )
-    },
-];
-
-const MAC_TEMPLATES: &[BackendTemplate] = &[
-    BackendTemplate {
-        default_ngl: Some("999"),
-        ..template(
-            "llama.cpp-mac",
-            "llama.cpp Metal",
-            "llama.cpp",
-            "llama.cpp-mac",
-            Some("llama-server"),
-            "llama.cpp Metal backend managed by OmniInfer",
-            &[
-                "chat",
-                "vision",
-                "stream",
-                "metal",
-                "apple",
-                "arm64",
-                "shared-memory",
-            ],
-            "OMNIINFER_LLAMA_CPP_MAC",
-        )
-    },
-    template(
-        "llama.cpp-mac-intel",
-        "llama.cpp macOS Intel",
-        "llama.cpp",
-        "llama.cpp-mac-intel",
-        Some("llama-server"),
-        "llama.cpp macOS Intel x64 backend managed by OmniInfer",
-        &["chat", "vision", "stream", "cpu", "macos", "x64", "intel"],
-        "OMNIINFER_LLAMA_CPP_MAC_INTEL",
-    ),
-    BackendTemplate {
-        default_ngl: Some("999"),
-        default_extra_args: &[
-            "-fa",
-            "on",
-            "--cache-type-k",
-            "turbo4",
-            "--cache-type-v",
-            "turbo4",
-        ],
-        log_file_name: "turboquant-server.log",
-        ..template(
-            "turboquant-mac",
-            "TurboQuant Metal",
-            "turboquant",
-            "turboquant-mac",
-            Some("llama-server"),
-            "TurboQuant llama.cpp-compatible Metal backend managed by OmniInfer on macOS",
-            &[
-                "chat",
-                "vision",
-                "stream",
-                "metal",
-                "apple",
-                "shared-memory",
-                "turboquant",
-            ],
-            "OMNIINFER_TURBOQUANT_MAC",
-        )
-    },
-    BackendTemplate {
-        runtime_mode: "embedded",
-        model_artifact: "directory",
-        supports_mmproj: false,
-        supports_ctx_size: false,
-        python_modules: &["mlx", "mlx_lm", "mlx_vlm", "torch", "torchvision"],
-        external_server_protocol: None,
-        ..template(
-            "mlx-mac",
-            "MLX LM/VLM",
-            "mlx-lm",
-            "mlx-mac",
-            None,
-            "Embedded MLX LM/VLM backend managed directly by OmniInfer on macOS",
-            &[
-                "chat",
-                "vision",
-                "stream",
-                "metal",
-                "apple",
-                "shared-memory",
-                "embedded",
-            ],
-            "OMNIINFER_MLX_MAC",
-        )
-    },
-];
-
-const ANDROID_TEMPLATES: &[BackendTemplate] = &[BackendTemplate {
-    default_ngl: Some("999"),
-    ..template(
-        "llama.cpp-android",
-        "llama.cpp Android",
-        "llama.cpp",
-        "llama.cpp-android",
-        Some("llama-server"),
-        "llama.cpp Android backend managed by OmniInfer",
-        &["chat", "vision", "stream", "android", "mobile"],
-        "OMNIINFER_LLAMA_CPP_ANDROID",
-    )
-}];
-
-const IOS_TEMPLATES: &[BackendTemplate] = &[
-    BackendTemplate {
-        default_ngl: Some("999"),
-        runtime_mode: "embedded",
-        external_server_protocol: None,
-        ..template(
-            "llama.cpp-ios",
-            "llama.cpp iOS",
-            "llama.cpp",
-            "llama.cpp-ios",
-            None,
-            "llama.cpp iOS Metal backend managed by OmniInfer",
-            &[
-                "chat", "vision", "stream", "metal", "apple", "mobile", "ios",
-            ],
-            "OMNIINFER_LLAMA_CPP_IOS",
-        )
-    },
-    BackendTemplate {
-        runtime_mode: "embedded",
-        model_artifact: "directory",
-        supports_mmproj: false,
-        supports_ctx_size: false,
-        external_server_protocol: None,
-        ..template(
-            "mlx-ios",
-            "MLX iOS",
-            "mlx-lm",
-            "mlx-ios",
-            None,
-            "Embedded MLX LM backend via mlx-swift on iOS",
-            &[
-                "chat", "stream", "metal", "apple", "mobile", "ios", "embedded",
-            ],
-            "OMNIINFER_MLX_IOS",
-        )
-    },
-];
 
 #[cfg(test)]
 mod tests {
