+++ crates/omniinfer-core/src/lib.rs
@@ -1,20 +1,27 @@
-pub mod anthropic;
-pub mod backend_args;
-pub mod backend_profiles;
-pub mod backend_registry;
-pub mod chat_stream;
-pub mod config;
-pub mod gateway_auth;
-pub mod http_client;
-pub mod local_state;
-pub mod model_artifacts;
-pub mod model_catalog;
-pub mod model_load;
-pub mod paths;
-pub mod public_models;
-pub mod request_normalization;
-pub mod resource_ledger;
-pub mod runtime_plan;
-pub mod runtime_process;
-pub mod serve_state;
-pub mod version;
+pub mod backend;
+pub mod model;
+pub mod protocol;
+pub mod runtime;
+pub mod state;
+pub mod support;
+
+pub use backend::args as backend_args;
+pub use backend::profiles as backend_profiles;
+pub use backend::registry as backend_registry;
+pub use model::artifacts as model_artifacts;
+pub use model::catalog as model_catalog;
+pub use model::load as model_load;
+pub use model::public as public_models;
+pub use protocol::anthropic;
+pub use protocol::chat_stream;
+pub use protocol::request_normalization;
+pub use runtime::plan as runtime_plan;
+pub use runtime::process as runtime_process;
+pub use runtime::resources as resource_ledger;
+pub use state::config;
+pub use state::local as local_state;
+pub use state::serve as serve_state;
+pub use support::gateway_auth;
+pub use support::http_client;
+pub use support::paths;
+pub use support::version;
