+++ crates/omniinfer-core/src/model/artifacts.rs
