+++ crates/omniinfer-core/src/model/load.rs
