+++ crates/omniinfer-core/src/model/mod.rs
@@ -1,0 +1,4 @@
+pub mod artifacts;
+pub mod catalog;
+pub mod load;
+pub mod public;
