+++ crates/omniinfer-core/src/model/public.rs
