+++ crates/omniinfer-core/src/protocol/anthropic.rs
