+++ crates/omniinfer-core/src/protocol/chat_stream.rs
