+++ crates/omniinfer-core/src/protocol/mod.rs
@@ -1,0 +1,3 @@
+pub mod anthropic;
+pub mod chat_stream;
+pub mod request_normalization;
