+++ crates/omniinfer-core/src/protocol/request_normalization.rs
