+++ crates/omniinfer-core/src/runtime/resources.rs
