+++ crates/omniinfer-core/src/state/local.rs
