+++ crates/omniinfer-core/src/state/mod.rs
@@ -1,0 +1,3 @@
+pub mod config;
+pub mod local;
+pub mod serve;
