+++ crates/omniinfer-core/src/state/serve.rs
