+++ crates/omniinfer-core/src/support/gateway_auth.rs
