+++ crates/omniinfer-core/src/support/mod.rs
@@ -1,0 +1,4 @@
+pub mod gateway_auth;
+pub mod http_client;
+pub mod paths;
+pub mod version;
